Fix decimal numbers in distance

+++ app/src/main/java/cz/cvut/fukalhan/repository/entity/Route.kt
@@ -3,6 +3,6 @@ package cz.cvut.fukalhan.repository.entity
 import com.google.android.gms.maps.model.LatLng
 
 data class Route(
-    var distance: Double = 0.0,
+    var distance: Int = 0,
     var newLocation: LatLng
 )
